backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pypdf2
pytesseract
pillow
sqlalchemy
requests
openai
pwdlib[bcrypt]
pwdlib[argon2]
python-dotenv
watchdog
apscheduler
opencv-python-headless
numpy
reportlab
python-jose[cryptography]
slowapi
python-magic-bin
